DAOS-4456 control: organise server instance files
In preparation for adding capability to manage IO server instances
independently, subdivide instance functionality between files.
