+++ src/control/server/instance_drpc.go
@@ -1,0 +1,77 @@
+//
+// (C) Copyright 2020 Intel Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// GOVERNMENT LICENSE RIGHTS-OPEN SOURCE SOFTWARE
+// The Government's rights to use, modify, reproduce, release, perform, display,
+// or disclose this software are subject to the terms of the Apache License as
+// provided in Contract No. 8F-30005.
+// Any reproduction of computer software, computer software documentation, or
+// portions thereof marked with this legend must also reproduce the markings.
+//
+
+package server
+
+import (
+	"github.com/golang/protobuf/proto"
+	"github.com/pkg/errors"
+
+	srvpb "github.com/daos-stack/daos/src/control/common/proto/srv"
+	"github.com/daos-stack/daos/src/control/drpc"
+)
+
+func (srv *IOServerInstance) setDrpcClient(c drpc.DomainSocketClient) {
+	srv.Lock()
+	defer srv.Unlock()
+	srv._drpcClient = c
+}
+
+func (srv *IOServerInstance) getDrpcClient() (drpc.DomainSocketClient, error) {
+	srv.RLock()
+	defer srv.RUnlock()
+	if srv._drpcClient == nil {
+		return nil, errors.New("no dRPC client set (data plane not started?)")
+	}
+	return srv._drpcClient, nil
+}
+
+// NotifyDrpcReady receives a ready message from the running IOServer
+// instance.
+func (srv *IOServerInstance) NotifyDrpcReady(msg *srvpb.NotifyReadyReq) {
+	srv.log.Debugf("%s instance %d drpc ready: %v", DataPlaneName, srv.Index(), msg)
+
+	// Activate the dRPC client connection to this iosrv
+	srv.setDrpcClient(drpc.NewClientConnection(msg.DrpcListenerSock))
+
+	go func() {
+		srv.drpcReady <- msg
+	}()
+}
+
+// AwaitDrpcReady returns a channel which receives a ready message
+// when the started IOServer instance indicates that it is
+// ready to receive dRPC messages.
+func (srv *IOServerInstance) AwaitDrpcReady() chan *srvpb.NotifyReadyReq {
+	return srv.drpcReady
+}
+
+// CallDrpc makes the supplied dRPC call via this instance's dRPC client.
+func (srv *IOServerInstance) CallDrpc(module, method int32, body proto.Message) (*drpc.Response, error) {
+	dc, err := srv.getDrpcClient()
+	if err != nil {
+		return nil, err
+	}
+
+	return makeDrpcCall(dc, module, method, body)
+}
